Решение конфликтов после rebase из cerberus-master-stream8
