--- src/main/java/school/faang/user_service/service/user/UserService.java
@@ -1,25 +1,0 @@
-package school.faang.user_service.service.user;
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Service;
-import school.faang.user_service.entity.User;
-import school.faang.user_service.repository.UserRepository;
-
-import java.util.List;
-import java.util.Optional;
-
-@Slf4j
-@RequiredArgsConstructor
-@Service
-public class UserService {
-    private final UserRepository userRepository;
-
-    public Optional<User> findUserById(Long userId) {
-        return userRepository.findById(userId);
-    }
-
-    public List<User> findAllUsers() {
-        return userRepository.findAll();
-    }
-}
