+++ src/test/java/school/faang/user_service/service/goal/GoalServiceTest.java
@@ -6,6 +6,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
+import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
 import school.faang.user_service.dto.goal.GoalFilterDto;
 import school.faang.user_service.dto.goal.RequestGoalDto;
@@ -16,13 +17,14 @@ import school.faang.user_service.entity.goal.GoalStatus;
 import school.faang.user_service.exeption.MaxActiveGoalsExceededException;
 import school.faang.user_service.exeption.NonExistentSkillException;
 import school.faang.user_service.repository.goal.GoalRepository;
+import school.faang.user_service.service.SkillService;
+import school.faang.user_service.service.UserService;
 import school.faang.user_service.service.filters.goal.GoalFilter;
-import school.faang.user_service.service.skill.SkillService;
-import school.faang.user_service.service.user.UserService;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.Optional;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -114,17 +116,15 @@ public class GoalServiceTest {
 
     @Test
     public void testCreateGoal_UserNotFound() {
-        when(userService.findUserById(1L)).thenReturn(Optional.empty());
-
-        EntityNotFoundException exception = assertThrows(EntityNotFoundException.class, () ->
-                goalService.createGoal(1L, goal1));
-
-        assertEquals("User not found", exception.getMessage());
+        when(userService.getUser(1L)).thenThrow(new NoSuchElementException("not found user with id 1"));
+        assertThrows(NoSuchElementException.class, () -> goalService.createGoal(1L, goal1),
+                "not found user with id 1");
+        Mockito.verify(userService, times(1)).getUser(1L);
     }
 
     @Test
     public void testCreateGoal_ExceedsActiveGoals() {
-        when(userService.findUserById(1L)).thenReturn(Optional.of(user));
+        when(userService.getUser(1L)).thenReturn(user);
         when(goalRepository.countActiveGoalsPerUser(1L)).thenReturn(4);
 
         MaxActiveGoalsExceededException exception = assertThrows(MaxActiveGoalsExceededException.class, () ->
@@ -135,7 +135,7 @@ public class GoalServiceTest {
 
     @Test
     public void testCreateGoal_NonExistentSkills() {
-        when(userService.findUserById(1L)).thenReturn(Optional.of(user));
+        when(userService.getUser(1L)).thenReturn(user);
         when(skillService.skillExistsByTitle("Skill")).thenReturn(false);
 
         NonExistentSkillException exception = assertThrows(NonExistentSkillException.class, () ->
@@ -146,7 +146,7 @@ public class GoalServiceTest {
 
     @Test
     public void testCreateGoal_Success() {
-        when(userService.findUserById(user.getId())).thenReturn(Optional.of(user));
+        when(userService.getUser(user.getId())).thenReturn(user);
         when(skillService.skillExistsByTitle("Skill")).thenReturn(true);
 
         goalService.createGoal(user.getId(), goal1);
